requirements.txt
Django==4.2.2
djangorestframework==3.14.0
psycopg2-binary==2.9.6
requests==2.31.0
django-redis==5.3.0
python-dotenv==1.0.0
pytest-django==4.5.2
responses==0.23.1